+++ src/main.jsx
@@ -3,6 +3,7 @@ import App from './App.jsx'
 import './index.css'
 import { installAuthFetch } from './lib/authFetch.js'
 import { initPwaInstall } from './lib/pwaInstall.js'
+import { I18nProvider } from './i18n'
 
 // Attach the remote-access token to API calls before anything renders.
 installAuthFetch()
@@ -10,7 +11,7 @@ installAuthFetch()
 initPwaInstall()
 
 createRoot(document.getElementById('root')).render(
-  <>
+  <I18nProvider>
     <App />
-  </>
+  </I18nProvider>
 )